Add templated memory utilities.
